[Fix] Fixes from @greymag
