+++ example/lib/tabs/banner_screen.dart
@@ -15,6 +15,7 @@ class _BannerScreenState extends State<BannerScreen> {
   late final AdStatusListener listener = AdStatusListener.banner(
       onAdLoaded: () => _changeAdStatus('Ad loaded'),
       onClickOnAD: () => _changeAdStatus('Clicked on ad'),
+      onShow: () => _changeAdStatus('Ad shown'),
       onNoAd: (reason) => _changeAdStatus('Ad not loaded: $reason'));
 
   @override
@@ -24,62 +25,21 @@ class _BannerScreenState extends State<BannerScreen> {
         return Column(
           children: [
             const SizedBox(height: 50),
-            InkWell(child: const Text('Create AD'), onTap: () async => onTapRewarded(context)
-                //  onTapInterstitial(context),
-                ),
             const SizedBox(height: 50),
-            InkWell(
-              child: const Text('Load AD'),
-              onTap: () async {},
+            SizedBox(
+              height: 50,
+              child: BannerAd(
+                widget._plugin,
+                794557,
+                listener: listener,
+              ),
             ),
-            const SizedBox(height: 50),
-            // InkWell(
-            //   child: const Text('Show AD'),
-            //   onTap: () async {
-            //     if (_ad == null) {
-            //       _showError(context, 'Ad not created');
-            //     } else {
-            //       _ad!.show();
-            //     }
-            //   },
-            // ),
-            SizedBox(height: 50, child: BannerAd(widget._plugin, '794557')),
             const SizedBox(height: 50),
             Text('Ad status: $_adStatus')
           ],
         );
       }),
     );
-  }
-
-  void _showError(BuildContext context, String text) {
-    showDialog(
-      context: context,
-      builder: (builder) {
-        return AlertDialog(
-          title: const Text('Error'),
-          content: Text(text),
-        );
-      },
-    );
-  }
-
-  Future<void> onTapRewarded(BuildContext context) async {
-    {
-      _ad = await widget._plugin.createBannerAd(45101);
-      // final adListener = AdStatusListener(
-      //     onAdLoaded: () => _changeAdStatus('Ad loaded'),
-      //     onDisplay: () => _changeAdStatus('Ad display'),
-      //     onClickOnAD: () => _changeAdStatus('Clicked on ad'),
-      //     onVideoCompleted: () => _changeAdStatus('Ad video completed'),
-      //     onDismiss: () => _changeAdStatus('Ad closed'),
-      //     onNoAd: (reason) => _changeAdStatus('Ad not loaded: $reason'));
-      // _ad?.addListener(adListener);
-      // _changeAdStatus('Ad created');
-
-      // await Future.sync(() => _ad?.load());
-      // _ad?.show();
-    }
   }
 
   void _changeAdStatus(String status) {
