+++ example/lib/tabs/interstitial_screen.dart
@@ -20,14 +20,7 @@ class _InterstitialScreenState extends State<InterstitialScreen> {
         return Column(
           children: [
             const SizedBox(height: 50),
-            InkWell(child: const Text('Create AD'), onTap: () async => onTapInterstitial(context)
-                //  onTapInterstitial(context),
-                ),
-            const SizedBox(height: 50),
-            InkWell(
-              child: const Text('Load AD'),
-              onTap: () async {},
-            ),
+            InkWell(child: const Text('Create AD'), onTap: () async => onTapInterstitial(context)),
             const SizedBox(height: 50),
             InkWell(
               child: const Text('Show AD'),
