+++ lib/my_target_flutter.dart
@@ -2,5 +2,3 @@ library my_target_flutter;
 
 export 'package:my_target_flutter/src/ad_status_listener.dart';
 export 'package:my_target_flutter/src/my_target_flutter.dart';
-
-export 'package:my_target_flutter/src/ad_types/types.dart';
