+++ lib/src/my_target_flutter.dart
@@ -1,12 +1,20 @@
 import 'dart:async';
+import 'dart:developer';
 
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
+import 'package:my_target_flutter/src/extensions.dart';
 
 import 'ad_status_listener.dart';
-import 'ad_types/base_interstitial_ad.dart';
 
-/// Implementing displaying MyTarget ads, currently supported by Interstitial Ads only.
+// Typedefs here for better user experience while using
+// this package
+
+typedef RewardedAd = _BaseInterstitialAd;
+typedef InterstitialAd = _BaseInterstitialAd;
+
+/// Implementing displaying MyTarget ads.
 class MyTargetFlutter {
   static const MethodChannel _channel = MethodChannel('my_target_flutter');
   static const channel = EventChannel('my_target_flutter/ad_listener');
@@ -24,7 +32,7 @@ class MyTargetFlutter {
 
   MyTargetFlutter({required this.isDebug});
 
-  Stream<AdEventMessage> get adListenStream => _stream ??= channel
+  Stream<AdEventMessage> get _adListenStream => _stream ??= channel
           .receiveBroadcastStream()
           .cast<Map<dynamic, dynamic>>()
           .transform(StreamTransformer.fromHandlers(
@@ -45,48 +53,54 @@ class MyTargetFlutter {
 
   /// Create an Interstitial ads with [slotId]
   Future<InterstitialAd> createInterstitialAd(int slotId) async {
-    final uid = await _channel.invokeMethod<String>(
-      _methodCreateInterstitialAd,
-      {'slotId': slotId},
-    );
-    if (uid == null) {
-      throw FlutterError('Can not create Interstitial ad');
-    } else {
-      return InterstitialAd(this, uid);
+    try {
+      final uid = await _channel.invokeMethod<String>(
+        _methodCreateInterstitialAd,
+        {'slotId': slotId},
+      );
+
+      assert(uid?.isNotEmpty ?? false);
+
+      return InterstitialAd(this, uid!);
+    } on Object catch (e, s) {
+      return Error.throwWithStackTrace(FlutterError('Can not create Interstitial ad'), s);
     }
   }
 
   /// Create an Rewarded ads with [slotId]
   Future<RewardedAd> createRewardedAd(int slotId) async {
-    final uid = await _channel.invokeMethod<String>(
-      _methodCreateRewardedAd,
-      {'slotId': slotId},
-    );
-    if (uid == null) {
-      throw FlutterError('Can not create Interstitial ad');
-    } else {
-      return RewardedAd(this, uid);
+    try {
+      final uid = await _channel.invokeMethod<String>(
+        _methodCreateRewardedAd,
+        {'slotId': slotId},
+      );
+
+      assert(uid?.isNotEmpty ?? false);
+
+      return RewardedAd(this, uid!);
+    } on Object catch (e, s) {
+      return Error.throwWithStackTrace(FlutterError('Can not create Rewarded ad'), s);
     }
   }
 
-  Future<dynamic> createBannerAd(int slotId) async {
-    final uid = await _channel.invokeMethod<String>(
-      _methodCreateBannerAd,
-      {'slotId': slotId},
-    );
-    print('BANNER COMES: $uid');
-    // if (uid == null) {
-    //   throw FlutterError('Can not create Interstitial ad');
-    // } else {
-    //   return RewardedAd(this, uid);
-    // }
+  Future createBannerAd(int slotId) async {
+    try {
+      final uid = await _channel.invokeMethod<String>(
+        _methodCreateBannerAd,
+        {'slotId': slotId},
+      );
+
+      assert(uid?.isNotEmpty ?? false);
+    } on Object catch (e, s) {
+      return Error.throwWithStackTrace(FlutterError('Can not create Banner ad'), s);
+    }
   }
 
-  Future<void> load(String uid) async {
+  Future<void> _load(String uid) async {
     await _channel.invokeMethod(_methodLoad, uid);
   }
 
-  Future<void> show(String uid) async {
+  Future<void> _show(String uid) async {
     await _channel.invokeMethod(_methodShow, uid);
   }
 
@@ -95,5 +109,106 @@ class MyTargetFlutter {
       'useDebugMode': useDebugMode,
       'testDevices': testDevices,
     };
+  }
+}
+
+/// Class for overlay advertisement like Rewarded Video & Interstitial
+///
+/// We using it cause myTargetSdk use similar logic under the hood
+///
+/// They're use BaseInterstitialAd class and extend it in
+/// InterstitialAd and RewardedAd classes
+///
+class _BaseInterstitialAd {
+  final MyTargetFlutter _plugin;
+  final String uid;
+
+  final _listeners = <AdStatusListener>{};
+
+  _BaseInterstitialAd(this._plugin, this.uid) {
+    _plugin._adListenStream.listen(_handleMessage);
+  }
+
+  Future<void> _handleMessage(AdEventMessage data) async {
+    if (data.uid == uid) {
+      _listeners.handleEvent(data);
+    }
+  }
+
+  void load() {
+    _plugin._load(uid);
+  }
+
+  void show() {
+    _plugin._show(uid);
+  }
+
+  void addListener(AdStatusListener listener) {
+    _listeners.add(listener);
+  }
+
+  void removeListener(AdStatusListener listener) {
+    _listeners.remove(listener);
+  }
+
+  void clearListeners() {
+    _listeners.clear();
+  }
+}
+
+/// Single banner advertisement
+class BannerAd extends StatefulWidget {
+  const BannerAd(this._plugin, this.uid, {Key? key, this.listener}) : super(key: key);
+  final MyTargetFlutter _plugin;
+  final int uid;
+  final AdStatusListener? listener;
+
+  @override
+  State<BannerAd> createState() => _BannerAdState();
+}
+
+class _BannerAdState extends State<BannerAd> {
+  final _listeners = <AdStatusListener>{};
+
+  @override
+  void initState() {
+    super.initState();
+    widget._plugin.createBannerAd(widget.uid);
+    if (widget.listener != null) _listeners.add(widget.listener!);
+    widget._plugin._adListenStream.listen(_handleMessage);
+  }
+
+  Future<void> _handleMessage(AdEventMessage data) async {
+    if (data.uid == widget.uid.toString()) {
+      _listeners.handleEvent(data);
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    const viewType = 'mytarget-banner-ad';
+
+    final creationParams = <String, dynamic>{
+      'id': widget.uid,
+    };
+
+    switch (defaultTargetPlatform) {
+      case TargetPlatform.android:
+        return AndroidView(
+          viewType: viewType,
+          layoutDirection: TextDirection.ltr,
+          creationParams: creationParams,
+          creationParamsCodec: const StandardMessageCodec(),
+        );
+      case TargetPlatform.iOS:
+        return UiKitView(
+          viewType: viewType,
+          layoutDirection: TextDirection.ltr,
+          creationParams: creationParams,
+          creationParamsCodec: const StandardMessageCodec(),
+        );
+      default:
+        throw UnsupportedError('Unsupported platform view');
+    }
   }
 }
